♻️ Refactor: chante the width
