+++ src/main/java/com/handongapp/cms/controller/v1/TbUserController.java
@@ -1,4 +1,4 @@
-package com.handongapp.cms.controller;
+package com.handongapp.cms.controller.v1;
 
 import com.handongapp.cms.dto.v1.TbUserDto;
 import com.handongapp.cms.security.PrincipalDetails;
